3000 BC Spawn cities always build a worker right away.
Civs may be a bit more aggressive in building Settlers and Workers.

Harappa is an AI civ now, though the 600 AD and 1700 AD scenarios have yet to be edited to portray the cities that they will usually settle.

Made some progress towards an AI Polynesia. Currently they will always settle their first two cities, even on the NW island which they would previously idle on. However, I can't seem to get any settlers that the AI builds to leave the city it's built in. It'll be loaded into a Waka, but the Waka never leaves.