AI POLYNESIA LIVES
Pyramids require a wonder in the city that it's built in. Egypt can rest easy knowing Babylon won't snipe the Pyramids while the work on the Sphinx, or vice versa.

It took a lot of fiddling with the AI and Settler Maps to get Polynesia to function as I wanted them to. They are still a work in progress, and I may need to tweak their AI in the future.

AI Polynesia has it's boat settler preplaced, otherwise it won't settle the eastern most core city

AI Polynesia gets a free Archer upon settling a city, otherwise it leaves many of it's cities undefended and collapses from unhappiness

AI Polynesia's settlermaps have been manhandled by my experimenting, every city spot that I feel is important is now worth 800. It seems to be working, oddly enough.

AI Polynesia has been taught to create fishing boats with their Wakas after offloading their settlers. Expect them to get up to speed much faster.

There seems to be an error regarding a great person and their name somewhere in the game, I'm not sure what's causing it though. If someone finds it please report it.