A single exclamation mark (!)
The AI became incompetently bad at the game when, a couple months ago, I absentmindedly placed an exclamation mark in the wrong place when rewriting a line of code.

I've been assuming the error was in the code I added, but apparently, for some stupid reason, I swapped the order of two variables, and brought an exclamation mark along with it.

Example if (not(true1 and true2))

Became if (true2 and not(true1))

I done messed up, and I was too focused on the changes that I made to notice I made a change without knowing it.

I love programming. There's something magical in how it's a testament to our abilty to make stupid mistakes and never notice until someone points it out.

Thank god for autocorrect.