Added Lower Egypt to 3000 BC. AI should be capable of conquering it.
When acquiring a city, all Free Buildings are automatically built in it. I would have had them be built after the Occupation ends so that said Occupation could represent a period of rebuilding, but it was too much work and I'm lazy.

Farms and Cottages take the same amount of time to build on Flood Plains as they do everywhere else.

Egypt is always at war with Independent2 if the latter controls Lower Egypt, this is done so that the AI prioritizes conquering it.

Fixed up placement of Lower Egypt in 1700 AD