HUGE NEW VERSION
Since the changes are so massive, the game will certainly be unbalanced and unstable. I'm releasing this version so that we can balance it and bugfix it together.

Huge rework of Game Speeds: Unit production is doubled and movement, attacks per turn, healing, and interceptions per turn have all been increased to proportionally match Marathon. This means 3 times the movement and 3 attacks per turn on Normal Speed and 1.5 times the movement and 2 attacks per turn on Epic.

Persia spawns in 550 AD and starts with more units. This aligns with the ascension of Cyrus to the throne and gives AI Babylon some more time to work towards capturing the Levant.

Persia starts with more units, including catapults.

Persia spawns at war with Egypt and Independent2 in addition to Babylonia and Carthage.

Ankuwash now builds Hulugannis, and very rapidly at that.

Added Djoser, a Ramesses lookalike that leads Egypt prior to building their first wonder.

The Egyptian and Babylonian AIs are much more warlike than ever before. Proper production management is crucial in the early game.

Added multiple scripted AI wars to help civs like Egypt and Babylonia conquer their historical territory. It is intended that they do not consistently succeed, the declaration of war may be scripted, but they alone decide how to wage it.

Tweaked Independent City ownership and starting units

Optimized Egyptian and Barbarian Huluganni movement AI. Don't expect them to wander into the Sahara, Rome or Russia anymore.

AI civs get 20% more production towards units and buildings, to make up for the fact that they aren't the smartest when it comes to what to build or what UnitAI it should have.